+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.3
 ---
 
+ * Added support for table prefix
  * Changed module icon
  * Improved internal structure